requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
black==23.9.1
click==8.1.7
crispy-bootstrap4==2022.1
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==2.3.0
pydantic_core==2.6.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
whitenoise==6.5.0
